pyproject.toml
[project]
name = "speech-normalizer"
version = "0.1.0"
description = "Robust speech volume normalizer using VAD-gated compression"
dependencies = [
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "numpy>=1.21.0",
    "webrtcvad-wheels>=2.0.10",
    "scipy>=1.9.0",
    "tqdm>=4.64.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
